refactor: moved skills to CharacterSkillsTrait
